requirements.txt
python-binance
pandas~=2.2.3
requests~=2.32.3
ta~=0.11.0
python-dotenv
psycopg2-binary  # для PostgreSQL